fill out the icons for inbox + separate addTask into component
